+++ CHANGELOG.md
@@ -5,6 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep A Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0-beta.1]
+
+### Added
+
+- "About Us" button now links to Tractornation.org
+
+### Fixed
+
+- Mobile UI issues
+- Qr codes not saving
+- Number input min and maximum
+- Small theme inconsistencies
+
+### Changed
+
+- Reworded the toggle for "Lead Scout Only"
+- Current device id now shows on appbar instead of version
+
 ## [0.2.0-beta]
 
 ### Added